requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
gunicorn==20.1.0
Pillow==9.5.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
